test 6 again

+++ data.c
@@ -3,7 +3,7 @@
 
 struct test{
   char first;
-  float second;
+  long second;
   int third;
   //char fourth;
   //char* fifth;
@@ -13,6 +13,6 @@ int main(int argc, char *argv[])
 {
   int size = sizeof(struct test);
   printf("Size is: %d \n", size);
-  printf("Size is: %d \n", (int)sizeof(float));
+  printf("Size is: %d \n", (int)sizeof(long));
   return 0;
 }
